% Auto-generated: do not edit by hand
\name{''DashAnnotationDropdown}

\alias{''DashAnnotationDropdown}

\title{DashAnnotationDropdown component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DashAnnotationDropdown(id=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Unnamed list. The data for the annotation dropdown as a dictionary}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

